Create asset report info widget
